* master: (50 commits)
  All teams are now "part" of every public contest and the API shows non-public contests if authenticated as a team or jury
  Fix example data: make one contest public and the other not. Also set deactivatetime when upgrading to something
  Fix problem shortnames.
  Improve error message when parsing contest starttime fails.
  Fix mysql example data.
  Uniform layout of SQL example data.
  Repair contest simulation script by readding option parsing and add option to specify contest (defaults to single active one).
  Move contest{problem,team} table to keep table list sorted.
  Various minor changes (formatting, comments, etc.) no functional changes intended.
  Don't use array_merge with numeric keys, since the numeric keys from the first array are overwritten by those of the second.
  Update changelog
  Allow linking of problems to contests
  Fix problem export
  Show more information in contest table of a problem
  Fix problem zip uploading
  Fix contest selection at problems + editing problems
  Don't need these hidden fields, as cookies take care of them
  Clarifications: fix it again and add a switch for showing only clarifications of currently selected contest
  Submission: show contest name and link to contest-specific team and problem pages
  Added an option to only show the balloons of the currently active contest
  ...

Conflicts:
	lib/lib.misc.php
	lib/www/checkers.jury.php
	lib/www/clarification.php
	lib/www/common.php
	lib/www/scoreboard.php
	sql/mysql_db_defaultdata.sql
	sql/mysql_db_structure.sql
	submit/submit.cc
	www/jury/clarification.php
	www/jury/contest.php
	www/jury/contests.php
	www/jury/submission.php
	www/team/clarification.php

+++ lib/www/common.jury.php
@@ -44,10 +44,24 @@ function editLink($table, $value, $multi = false)
  */
 function delLink($table, $field, $value)
 {
-	return "<a href=\"delete.php?table=" . urlencode($table) . "&amp;" .
-		$field . "=" . urlencode($value) ."\"><img src=\"../images/delete.png\" " .
-		"alt=\"delete\" title=\"delete this " . htmlspecialchars($table) .
-		"\" class=\"picto\" /></a>";
+	return delLinkMultiple($table, array($field), array($value));
+}
+
+/**
+ * Return a link to delete a specific data element from a given table.
+ * Takes the table, the key fields to match on and the values.
+ */
+function delLinkMultiple($table, $fields, $values, $referrer = '')
+{
+	$arguments = '';
+	foreach ($fields as $i => $field) {
+		$arguments .= '&amp;' . $field . '=' . urlencode($values[$i]);
+	}
+	return "<a href=\"delete.php?table=" . urlencode($table) . $arguments .
+	       "&amp;referrer=" . urlencode($referrer)
+	       ."\"><img src=\"../images/delete.png\" " .
+	       "alt=\"delete\" title=\"delete this " . htmlspecialchars($table) .
+	       "\" class=\"picto\" /></a>";
 }
 
 /**
@@ -161,13 +175,13 @@ if (!function_exists('parse_ini_string')) {
  * and update problem with it, or insert new problem when probid=NULL.
  * Returns probid on success, or generates error on failure.
  */
-function importZippedProblem($zip, $probid = NULL)
+function importZippedProblem($zip, $probid = NULL, $cid = -1)
 {
 	global $DB, $teamid;
 	$prop_file = 'domjudge-problem.ini';
 
-	$ini_keys = array('probid', 'cid', 'name', 'allow_submit', 'allow_judge',
-	                  'timelimit', 'special_run', 'special_compare', 'color');
+	$ini_keys_problem = array('name', 'timelimit', 'special_run', 'special_compare');
+	$ini_keys_contest_problem = array('probid', 'allow_submit', 'allow_judge', 'color');
 
 	$def_timelimit = 10;
 
@@ -180,31 +194,50 @@ function importZippedProblem($zip, $probid = NULL)
 		}
 	} else {
 		// Only preserve valid keys:
-		$ini_array = array_intersect_key($ini_array,array_flip($ini_keys));
+		$ini_array_problem = array_intersect_key($ini_array,array_flip($ini_keys_problem));
+		$ini_array_contest_problem = array_intersect_key($ini_array,array_flip($ini_keys_contest_problem));
 
 		if ( $probid===NULL ) {
-			if ( !isset($ini_array['probid']) ) {
+			if ( !isset($ini_array_contest_problem['probid']) ) {
 				error("Need 'probid' in '" . $prop_file . "' when adding a new problem.");
 			}
-			// Set sensible defaults for cid and name if not specified:
-			if ( !isset($ini_array['cid'])       ) $ini_array['cid'] = getCurContest();
-			if ( !isset($ini_array['name'])      ) $ini_array['name'] = $ini_array['probid'];
-			if ( !isset($ini_array['timelimit']) ) $ini_array['timelimit'] = $def_timelimit;
+			// Set sensible defaults for name and timelimit if not specified:
+			if ( !isset($ini_array_problem['name'])      ) $ini_array_problem['name'] = $ini_array_problem['probid'];
+			if ( !isset($ini_array_problem['timelimit']) ) $ini_array_problem['timelimit'] = $def_timelimit;
 
 			// rename probid to shortname
-			$probid = $ini_array['probid'];
-			unset($ini_array['probid']);
-			$ini_array['shortname'] = $probid;
+			$shortname = $ini_array_contest_problem['probid'];
+			unset($ini_array_contest_problem['probid']);
+			$ini_array_contest_problem['shortname'] = $shortname;
 
-			$probid = $DB->q('RETURNID INSERT INTO problem (' .
-			                 implode(', ',array_keys($ini_array)) .
-			                 ') VALUES (%As)', $ini_array);
+			$probid = $DB->q('RETURNID INSERT INTO problem (' . implode(', ',array_keys($ini_array_problem)) .
+			       ') VALUES %As', $ini_array_problem);
+
+			if ($cid != -1) {
+				$ini_array_contest_problem['cid'] = $cid;
+				$ini_array_contest_problem['probid'] = $probid;
+				$DB->q('INSERT INTO contestproblem (' . implode(', ',array_keys($ini_array_contest_problem)) .
+				       ') VALUES %As', $ini_array_contest_problem);
+			}
 		} else {
-			// Remove keys that cannot be modified:
-			unset($ini_array['probid']);
-			unset($ini_array['cid']);
 
-			$DB->q('UPDATE problem SET %S WHERE probid = %i', $ini_array, $probid);
+			$DB->q('UPDATE problem SET %S WHERE probid = %i', $ini_array_problem, $probid);
+
+			if ( $cid != -1 ) {
+				if ($DB->q("MAYBEVALUE SELECT probid FROM contestproblem WHERE probid = %i AND cid = %i", $probid, $cid)) {
+					// Remove keys that cannot be modified:
+					unset($ini_array_contest_problem['probid']);
+					$DB->q('UPDATE contestproblem SET %S WHERE probid = %i AND cid = %i', $ini_array_contest_problem, $probid, $cid);
+				} else {
+					$shortname = $ini_array_contest_problem['probid'];
+					unset($ini_array_contest_problem['probid']);
+					$ini_array_contest_problem['shortname'] = $shortname;
+					$ini_array_contest_problem['cid'] = $cid;
+					$ini_array_contest_problem['probid'] = $probid;
+					$DB->q('INSERT INTO contestproblem (' . implode(', ',array_keys($ini_array_contest_problem)) .
+					       ') VALUES %As', $ini_array_contest_problem);
+				}
+			}
 		}
 	}
 
@@ -247,8 +280,7 @@ function importZippedProblem($zip, $probid = NULL)
 	echo "</ul>\n<p>Added $ncases testcase(s).</p>\n";
 
 	// submit reference solutions
-	if ( $DB->q('VALUE SELECT allow_submit FROM problem
-	             WHERE probid = %i', $probid) && !empty($teamid) ) {
+	if ( $cid != -1 && $DB->q('VALUE SELECT allow_submit FROM problem INNER JOIN contestproblem using (probid) WHERE probid = %i AND cid = %i', $probid, $cid) ) {
 		// First find all submittable languages:
 		$langs = $DB->q('KEYVALUETABLE SELECT langid, extensions
  		                 FROM language WHERE allow_submit = 1');
@@ -266,17 +298,14 @@ function importZippedProblem($zip, $probid = NULL)
 					break;
 				}
 			}
-			if ( !empty($langid) ) {
+			if( !empty($langid) && !empty($teamid) ) {
 				if ( !($tmpfname = tempnam(TMPDIR, "ref_solution-")) ) {
 					error("Could not create temporary file.");
 				}
 				file_put_contents($tmpfname, $zip->getFromIndex($j));
 				if( filesize($tmpfname) <= dbconfig_get('sourcesize_limit')*1024 ) {
-					submit_solution($teamid, $probid, $langid, array($tmpfname), array($filename));
-					echo "<li>Added jury solution from: <tt>$filename</tt></li>\n";
-					$njurysols++;
-				} else {
-					echo "<li>Could not add jury solution <tt>$filename</tt>: too large.</li>\n";
+					submit_solution($teamid, $probid, $cid, $langid,
+							array($tmpfname), array($filename));
 				}
 				unlink($tmpfname);
 			}
